chore: re-organized scripts workspace sources structure
