+++ scripts/src/constants.ts
@@ -1,0 +1,1 @@
+export const PLUGIN_NAME_PREFIX = '@omh/react-native-maps-plugin-';
