+++ scripts/src/templates.ts
